platform introspection in a Process()
