+++ pants-plugins/uses_services/platform.py
@@ -1,39 +1,35 @@
-import platform
-from dataclasses import dataclass
+import json
 
-import distro
-
+from pants.engine.fs import EMPTY_DIGEST, CreateDigest, Digest, FileContent, FileDigest
+from pants.engine.process import Process, ProcessCacheScope, ProcessResult
 from pants.engine.rules import collect_rules, _uncacheable_rule
+from .inspect_platform import Platform
 
-
-@dataclass(frozen=True)
-class Platform:
-    arch: str
-    os: str
-    distro: str
-    distro_name: str
-    distro_codename: str
-    distro_like: str
-    distro_major_version: str
-    distro_version: str
-    mac_release: str
-    win_release: str
+__all__ = ["Platform", "get_platform", "rules"]
 
 
 @_uncacheable_rule
 async def get_platform() -> Platform:
-    return Platform(
-        arch=platform.machine(),  # x86_64
-        os=platform.system(),  # Linux, Darwin
-        distro=distro.id(),  # rhel, ubuntu, centos, gentoo, darwin
-        distro_name=distro.name(),  # Ubuntu, Centos Linux, Gentoo, Darwin
-        distro_codename=distro.codename(),  # xenial, Core, n/a, ''
-        distro_like=distro.like(),  # debian, rhel fedora, '', ''
-        distro_major_version=distro.major_version(),  # 16, 7, 2, 19
-        distro_version=distro.version(),  # 16.04, 7, 2.7, 19.6.0
-        mac_release=platform.mac_ver()[0],  # '', 10.15.7
-        win_release=platform.win32_ver()[0],  # ''
+
+    script_path = "./inspect_platform.py"
+    script_digest = await Get(
+        Digest,
+        # TODO: I need both the script and `distro`, the dependency. Do I need to build a PEX?
+        CreateDigest([FileContent(script_path, "", is_executable=True)]),
     )
+
+    result = await Get(
+        ProcessResult,
+        Process(
+            description=f"Introspecting platform (arch, os, distro)",
+            input_digest=script_digest,
+            argv=[script_path],
+            # this can change from run to run, so don't cache results.
+            cache_scope=ProcessCacheScope.PER_RESTART,  # NEVER?
+        ),
+    )
+    platform = json.loads(result.stdout)
+    return Platform(**platform)
 
 
 def rules():
